Move premium tab to the new settings
[MAILPOET-2681]
