--- assets/js/src/settings/pages/key_activation.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function KeyActivation() {
-  return <p>Key Activation!</p>;
-}
