+++ assets/js/src/settings/pages/key_activation/messages/key_messages.jsx
