+++ assets/js/src/settings/pages/key_activation/messages/mss_messages.jsx
