+++ assets/js/src/settings/pages/key_activation/messages/premium_installation_messages.jsx
